Keep revamping facet-reflect API for increased safety
